requirements.txt
streamlit
requests
flask
python-dotenv
beautifulsoup4
pydantic
langchain_core
langchain_groq
gtts
httpx
groq
translate